+++ pyccel/codegen/wrapper/fortran_to_c_wrapper.py
@@ -281,7 +281,7 @@ class FortranToCWrapper(Wrapper):
 
         self.exit_scope()
 
-        func = BindCFunctionDef(name, func_arguments, func_results, body, scope=func_scope, original_function = expr,
+        func = BindCFunctionDef(name, func_arguments, body, func_results, scope=func_scope, original_function = expr,
                 docstring = expr.docstring, result_pointer_map = expr.result_pointer_map)
 
         self.scope.functions[name] = func
@@ -563,7 +563,7 @@ class FortranToCWrapper(Wrapper):
         self._additional_exprs.clear()
         self.exit_scope()
 
-        getter = BindCFunctionDef(getter_name, (getter_arg,), (getter_result,), getter_body,
+        getter = BindCFunctionDef(getter_name, (getter_arg,), getter_body, (getter_result,),
                                 original_function = expr, scope = getter_scope)
 
         # ----------------------------------------------------------------------------------
@@ -601,7 +601,7 @@ class FortranToCWrapper(Wrapper):
             setter_body.append(Assign(attrib, set_val))
         self.exit_scope()
 
-        setter = BindCFunctionDef(setter_name, setter_args, (), setter_body,
+        setter = BindCFunctionDef(setter_name, setter_args, setter_body,
                                 original_function = expr, scope = setter_scope)
         return BindCClassProperty(lhs.cls_base.scope.get_python_name(expr.name),
                                   getter, setter, lhs.dtype)
@@ -645,7 +645,7 @@ class FortranToCWrapper(Wrapper):
         c_loc = CLocFunc(local_var, bind_var)
         body = [alloc, c_loc]
 
-        new_method = BindCFunctionDef(func_name, [], [result], body, original_function = None, scope = func_scope)
+        new_method = BindCFunctionDef(func_name, [], body, [result], original_function = None, scope = func_scope)
 
         methods = [self._wrap(m) for m in expr.methods]
         methods = [m for m in methods if not isinstance(m, EmptyNode)]