+++ pyccel/stdlib/STC_Extensions/Set_extensions.h
@@ -1,5 +1,3 @@
-#ifndef SET_EXTENSIONS_H
-#define SET_EXTENSIONS_H
 #include <stdarg.h>
 
 // This function represents a call to the .pop() method.
@@ -43,4 +41,3 @@ static inline i_type _c_MEMB(_union)(i_type* self, int n, ...) {
 #undef i_type
 #undef i_key
 #include <stc/priv/template2.h>
-#endif